refactor: rename base.ts
